cliente controller

+++ src/main/java/com/example/hamacasbackend/entidades/cliente/Cliente.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idCliente;
+    private Long idCliente;
 
     private String nombreCompleto;
